==> Code/00_setup_packges.r <==
### install required packages

if (!require("bibliometrix")) install.packages("bibliometrix")
if (!require("mice")) install.packages("mice")
if (!require("readxl")) install.packages("readxl")
if (!require("stargazer")) install.packages("stargazer")
if (!require("tidytext")) install.packages("tidytext")
if (!require("wordcloud")) install.packages("wordcloud")
if (!require("writexl")) install.packages("writexl")
